requirements.txt
certifi==2020.6.20
chardet==3.0.4
cryptowatch-sdk==0.0.12
idna==2.10
marshmallow==3.7.0
netifaces==0.10.6
protobuf==3.12.2
python-engineio==3.13.1
python-socketio==4.6.0
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
urllib3==1.25.9
websocket-client==0.57.0
